Renamed component Blog to Blogs
