+++ app/javascript/src/components/Blogs/SearchFilterPan.jsx
